+++ randomgen/__init__.py
@@ -9,14 +9,14 @@ from randomgen.threefry import ThreeFry
 from randomgen.threefry32 import ThreeFry32
 from randomgen.xoroshiro128 import Xoroshiro128
 from randomgen.xorshift1024 import Xorshift1024
-from randomgen.xoshiro256starstar import Xoshiro256StarStar
-from randomgen.xoshiro512starstar import Xoshiro512StarStar
+from randomgen.xoshiro256 import Xoshiro256
+from randomgen.xoshiro512 import Xoshiro512
 
 from ._version import get_versions
 
 __all__ = ['Generator', 'DSFMT', 'MT19937', 'PCG64', 'PCG32', 'Philox',
            'ThreeFry', 'ThreeFry32', 'Xoroshiro128', 'Xorshift1024',
-           'Xoshiro256StarStar', 'Xoshiro512StarStar', 'RandomState']
+           'Xoshiro256', 'Xoshiro512', 'RandomState']
 
 
 __version__ = get_versions()['version']