+++ randomgen/src/xoshiro256/xoshiro256-test-data-gen.c
@@ -1,13 +1,13 @@
 /*
  * Generate testing csv files
  *
- *  cl xoshiro512starstar-test-data-gen.c xoshiro512starstar.orig.c /
+ *  cl xoshiro256-test-data-gen.c xoshiro256.orig.c /
  * ../splitmix64/splitmix64.c /Ox
- * xoshiro512starstar-test-data-gen.exe *
+ * xoshiro256-test-data-gen.exe *
  *
- *  gcc xoshiro512starstar-test-data-gen.c xoshiro512starstar.orig.c /
- * ../splitmix64/splitmix64.c -o xoshiro512starstar-test-data-gen
- *  ./xoshiro512starstar-test-data-gen
+ *  gcc xoshiro256-test-data-gen.c xoshiro256.orig.c /
+ * ../splitmix64/splitmix64.c -o xoshiro256-test-data-gen
+ *  ./xoshiro256-test-data-gen
  *
  * Requres the Random123 directory containing header files to be located in the
  * same directory (not included).
@@ -15,7 +15,7 @@
  */
 
 #include "../splitmix64/splitmix64.h"
-#include "xoshiro512starstar.orig.h"
+#include "xoshiro256.orig.h"
 #include <inttypes.h>
 #include <stdio.h>
 
@@ -26,7 +26,7 @@ int main() {
   uint64_t state, seed = 0xDEADBEAF;
   state = seed;
   int i;
-  for (i = 0; i < 8; i++) {
+  for (i = 0; i < 4; i++) {
     s[i] = splitmix64_next(&state);
   }
   uint64_t store[N];
@@ -35,7 +35,7 @@ int main() {
   }
 
   FILE *fp;
-  fp = fopen("xoshiro512starstar-testset-1.csv", "w");
+  fp = fopen("xoshiro256-testset-1.csv", "w");
   if (fp == NULL) {
     printf("Couldn't open file\n");
     return -1;
@@ -50,13 +50,13 @@ int main() {
   fclose(fp);
 
   seed = state = 0;
-  for (i = 0; i < 8; i++) {
+  for (i = 0; i < 4; i++) {
     s[i] = splitmix64_next(&state);
   }
   for (i = 0; i < N; i++) {
     store[i] = next();
   }
-  fp = fopen("xoshiro512starstar-testset-2.csv", "w");
+  fp = fopen("xoshiro256-testset-2.csv", "w");
   if (fp == NULL) {
     printf("Couldn't open file\n");
     return -1;