+++ randomgen/xoshiro512starstar.py
@@ -1,0 +1,11 @@
+from randomgen.xoshiro512 import Xoshiro512
+import warnings
+
+
+def Xoshiro512StarStar(*args, **kwargs):
+    """
+    This is a deprecation shim.  Use Xoshiro512
+    """
+    warnings.warn('Xoshiro512StarStar has been renamed Xoshiro512', 
+                  DeprecationWarning)    
+    return Xoshiro512(*args, **kwargs)